SimpleLogger was renamed to FileSimpleLogger.
